Issue fixed with Rental model

+++ Vidly/Models/Rental.cs
@@ -10,7 +10,7 @@ namespace Vidly.Models
         public Customer Customer { get; set; }
         
         [Required]
-        public long CustomerId { get; set; }
+        public int CustomerId { get; set; }
         
         public Movie Movie { get; set; }
